- 0.5.0初次提交 - 结构优化，大部分早期代码调整到java8语言规范 - SingletonMap改为Builder模式

+++ coodex-libs-pom/coodex-utilities/src/main/java/org/coodex/concurrent/Debouncer.java
@@ -24,11 +24,11 @@ import java.util.concurrent.TimeUnit;
 public class Debouncer<T> extends AbstractCoalition<T> {
     private ScheduledFuture prevFuture = null;
 
-    public Debouncer(Callback<T> c, int interval, ScheduledExecutorService scheduledExecutorService) {
+    public Debouncer(Callback<T> c, long interval, ScheduledExecutorService scheduledExecutorService) {
         super(c, interval, scheduledExecutorService);
     }
 
-    public Debouncer(Callback<T> c, int interval) {
+    public Debouncer(Callback<T> c, long interval) {
         super(c, interval);
     }
